+++ tests/Functional/Resources/Transactions/TransactionsClientTest.php
@@ -11,6 +11,7 @@ use Paddle\SDK\Entities\Shared\BillingDetails;
 use Paddle\SDK\Entities\Shared\CollectionMode;
 use Paddle\SDK\Entities\Shared\CurrencyCode;
 use Paddle\SDK\Entities\Shared\CustomData;
+use Paddle\SDK\Entities\Shared\Disposition;
 use Paddle\SDK\Entities\Shared\Interval;
 use Paddle\SDK\Entities\Shared\Money;
 use Paddle\SDK\Entities\Shared\PriceQuantity;
@@ -529,11 +530,13 @@ class TransactionsClientTest extends TestCase
      * @dataProvider getInvoicePDFOperationsProvider
      */
     public function get_invoice_pdf_hits_expected_uri(
+        string $id,
+        Disposition|null $disposition,
         ResponseInterface $response,
         string $expectedUri,
     ): void {
         $this->mockClient->addResponse($response);
-        $this->client->transactions->getInvoicePDF('txn_01hen7bxc1p8ep4yk7n5jbzk9r');
+        $this->client->transactions->getInvoicePDF($id, $disposition);
         $request = $this->mockClient->getLastRequest();
 
         self::assertInstanceOf(RequestInterface::class, $request);
@@ -544,8 +547,24 @@ class TransactionsClientTest extends TestCase
     public static function getInvoicePDFOperationsProvider(): \Generator
     {
         yield 'Default' => [
+            'txn_01hen7bxc1p8ep4yk7n5jbzk9r',
+            null,
             new Response(200, body: self::readRawJsonFixture('response/get_invoice_pdf_default')),
             sprintf('%s/transactions/txn_01hen7bxc1p8ep4yk7n5jbzk9r/invoice', Environment::SANDBOX->baseUrl()),
+        ];
+
+        yield 'Disposition Inline' => [
+            'txn_02hen7bxc1p8ep4yk7n5jbzk9r',
+            Disposition::Inline(),
+            new Response(200, body: self::readRawJsonFixture('response/get_invoice_pdf_default')),
+            sprintf('%s/transactions/txn_02hen7bxc1p8ep4yk7n5jbzk9r/invoice?disposition=inline', Environment::SANDBOX->baseUrl()),
+        ];
+
+        yield 'Disposition Attachment' => [
+            'txn_03hen7bxc1p8ep4yk7n5jbzk9r',
+            Disposition::Attachment(),
+            new Response(200, body: self::readRawJsonFixture('response/get_invoice_pdf_default')),
+            sprintf('%s/transactions/txn_03hen7bxc1p8ep4yk7n5jbzk9r/invoice?disposition=attachment', Environment::SANDBOX->baseUrl()),
         ];
     }
 }